add copyright ant fix import
