+++ ci/requirements-wheel.txt
@@ -7,4 +7,3 @@ scipy==1.5.4; python_version>='3.9'
 cython==0.29.21
 requests>=2.4.0
 cloudpickle>=1.5.0
-oss2
