+++ mars/lib/filesystem/_oss_lib/common.py
@@ -1,13 +1,29 @@
-import os
-from functools import lru_cache
-from configparser import ConfigParser
-from pathlib import Path
-from urllib import parse
-import oss2
-import json
+# Copyright 1999-2021 Alibaba Group Holding Ltd.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import base64
+import json
+import os
+
+from .import_err_handler import ImportErrorHandler
 from ..base import path_type, stringify_path, FileEntry
 
+try:
+	import oss2
+except ImportError:
+	oss2 = ImportErrorHandler('oss2')
+	
 # OSS api time out
 _oss_time_out = 10
 
@@ -170,21 +186,6 @@ def oss_scandir(dirname: path_type):
                     "modified_time": obj.last_modified,
                 }
             )
-
-
-def convert_oss_path(path: path_type, access_key_id, access_key_secret, end_point):
-	"""
-	Used to register the access_key_id, access_key_secret and endpoint of OSS.
-	The access_key_id and endpoint are put into the url with url-safe-base64 encoding.
-	"""
-	if isinstance(path, (list, tuple)):
-		path = path[0]
-	param_dict = {'access_key_id': access_key_id, 'end_point': end_point}
-	id_endpoint = dict_to_url(param_dict)
-	password = access_key_secret
-	parse_result = parse.urlparse(path)
-	new_path = f'{parse_result.scheme}://{id_endpoint}:{password}@{parse_result.netloc}{parse_result.path}'
-	return new_path
 
 
 def dict_to_url(param: dict):
