+++ mars/lib/filesystem/tests/test_oss.py
@@ -1,10 +1,25 @@
-import mock
-import oss2  # 2.14.0
-from mars.lib.filesystem import oss
-from mars.lib.filesystem._oss_lib.common import convert_oss_path
+# Copyright 1999-2021 Alibaba Group Holding Ltd.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import time
 from io import BytesIO
+
+import mock
+
 import mars.dataframe as md
+from mars.lib.filesystem import oss
+from mars.lib.filesystem.oss import convert_oss_path
 
 
 class oss_obj_info:
